mcp-phi-base/workers/validation/requirements.txt
# Validation Worker Dependencies
numpy==1.24.3
scipy==1.10.1
sympy==1.12
mpmath==1.3.0  # High precision math
z3-solver==4.12.2.0  # SMT solver for theorem proving
pika==1.3.2
redis==4.5.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0
structlog==23.1.0